readme.md
my first app. cross compatible.
I created it because I am so sick of trying to remember a million passwords and it is intended to run from a USB stick so that it can be used on multiple devices.
has a long to do list that I have no further plans to develop for the time being.